simplify everything

+++ src/main/java/UEVEZon/models/Produit.java
@@ -4,9 +4,9 @@ import java.util.Date;
 
 public class Produit {
     public Date entree;
-    public Long serie;
+    public Integer serie;
 
-    public Produit(Long identifiantProduit) {
+    public Produit(Integer identifiantProduit) {
         this.serie = identifiantProduit;
     }
 }
